question-generator/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
PyPDF2==3.0.1
google-generativeai==0.8.5
python-dotenv==1.1.1
pydantic==2.11.7
reportlab==4.4.3
requests==2.32.4
PyJWT==2.10.1
streamlit==1.28.1